requirements.txt
# Core App
streamlit==1.32.0
python-dotenv==1.0.1

# PDF Reading
PyPDF2==3.0.1

# LangChain & Google Gemini
langchain==0.1.12
langchain-community>=0.0.28,<0.1
google-generativeai==0.3.2

# Vector Store
chromadb==0.4.24

# Image support
Pillow==10.0.0

# Added from the code block
sentence-transformers==2.5.1
pypdf==4.1.0

# Added from the code block
python-multipart==0.0.9

# Added from the code block
numpy>=1.24.0,<2.0.0
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
